+++ auth0_flutter/ios/Classes/AuthAPI/AuthAPIHandler.swift
@@ -11,7 +11,8 @@ typealias AuthAPIMethodHandlerProvider = (_ method: AuthAPIHandler.Method, _ cli
 public class AuthAPIHandler: NSObject, FlutterPlugin {
     enum Method: String, CaseIterable {
         case loginWithUsernameOrEmail = "auth#login"
-        case loginWithOtp = "auth#loginOtp"
+        case loginWithOTP = "auth#loginOtp"
+        case multifactorChallenge = "auth#multifactorChallenge"
         case signup = "auth#signUp"
         case userInfo = "auth#userInfo"
         case renew = "auth#renew"
@@ -36,7 +37,8 @@ public class AuthAPIHandler: NSObject, FlutterPlugin {
     var methodHandlerProvider: AuthAPIMethodHandlerProvider = { method, client in
         switch method {
         case .loginWithUsernameOrEmail: return AuthAPILoginUsernameOrEmailMethodHandler(client: client)
-        case .loginWithOtp: return AuthAPILoginWithOtpMethodHandler(client: client)
+        case .loginWithOTP: return AuthAPILoginWithOTPMethodHandler(client: client)
+        case .multifactorChallenge: return AuthAPIMultifactorChallengeMethodHandler(client: client)
         case .signup: return AuthAPISignupMethodHandler(client: client)
         case .userInfo: return AuthAPIUserInfoMethodHandler(client: client)
         case .renew: return AuthAPIRenewMethodHandler(client: client)