+++ auth0_flutter_platform_interface/lib/src/method_channel_auth0_flutter_auth.dart
@@ -3,10 +3,12 @@ import 'package:flutter/services.dart';
 import 'auth/api_exception.dart';
 import 'auth/auth_login_options.dart';
 import 'auth/auth_login_with_otp_options.dart';
+import 'auth/auth_multifactor_challenge_options.dart';
 import 'auth/auth_renew_access_token_options.dart';
 import 'auth/auth_reset_password_options.dart';
 import 'auth/auth_signup_options.dart';
 import 'auth/auth_user_info_options.dart';
+import 'auth/challenge.dart';
 import 'auth0_flutter_auth_platform.dart';
 import 'credentials.dart';
 import 'database_user.dart';
@@ -17,6 +19,7 @@ import 'user_profile.dart';
 const MethodChannel _channel = MethodChannel('auth0.com/auth0_flutter/auth');
 const String authLoginMethod = 'auth#login';
 const String authLoginWithOtpMethod = 'auth#loginOtp';
+const String authMultifactorChallengeMethod = 'auth#multifactorChallenge';
 const String authUserInfoMethod = 'auth#userInfo';
 const String authSignUpMethod = 'auth#signUp';
 const String authRenewMethod = 'auth#renew';
@@ -38,6 +41,15 @@ class MethodChannelAuth0FlutterAuth extends Auth0FlutterAuthPlatform {
         await invokeRequest(method: authLoginWithOtpMethod, request: request);
 
     return Credentials.fromMap(result);
+  }
+
+  @override
+  Future<Challenge> multifactorChallenge(
+      final ApiRequest<AuthMultifactorChallengeOptions> request) async {
+    final Map<String, dynamic> result = await invokeRequest(
+        method: authMultifactorChallengeMethod, request: request);
+
+    return Challenge.fromMap(result);
   }
 
   @override